Demonstrate the Gtk-Message
When Open file.. from the File menu is selected on Linux Mint 19.2
(Tina) with Node.js 12.8.1, Chromium 78.0.3904.130, and Electron 7.1.10,
the following message appears on the main process console:

Gtk-Message: 21:56:24.820: GtkDialog mapped without a transient parent. This is discouraged.

This was reported in electron issue #12423 but the issue was closed due
to lack of response to request for a fork of electron-quick-start after
ckerr reported that he could no longer reproduce. This is in response
and perhaps allows the issue to be investigated further.